Format and add doxygen

+++ src/Source_Files/SHTC3.c
@@ -59,32 +59,33 @@ void shtc3_i2c_begin_read(scheduler_event_t cb_event) {
   i2c_write(shtc3_i2c_x, SHTC3_ADDR, cmd_sleep, SHTC_CMD_LEN, 0);
 }
 
-//uint8_t crc8(uint8_t d[], uint32_t nBytes, uint8_t poly) {
-//  uint8_t remainder = 0;
-//  for (uint32_t byte = 0; byte < nBytes; ++byte) {
-//    remainder ^= d[byte];
-//    for (uint32_t bit = 8; bit > 0; --bit) {
-//      if (remainder & 0x8000)
-//        remainder = (remainder << 1) ^ poly;
-//      else
-//        remainder = (remainder << 1);
-//    }
-//  }
-//  return remainder;
-//}
+// uint8_t crc8(uint8_t d[], uint32_t nBytes, uint8_t poly) {
+//   uint8_t remainder = 0;
+//   for (uint32_t byte = 0; byte < nBytes; ++byte) {
+//     remainder ^= d[byte];
+//     for (uint32_t bit = 8; bit > 0; --bit) {
+//       if (remainder & 0x8000)
+//         remainder = (remainder << 1) ^ poly;
+//       else
+//         remainder = (remainder << 1);
+//     }
+//   }
+//   return remainder;
+// }
 
 bool shtc3_get_temperature_and_humidity(float *temperature, float *humidity) {
   EFM_ASSERT(shtc3_module_open);
   // Parse packet buffer and compute CRC
-  uint16_t temp_raw     = (shtc3_rx_buff[0] << 8) | shtc3_rx_buff[1];
-  uint16_t hum_raw      = (shtc3_rx_buff[3] << 8) | shtc3_rx_buff[4];
-//  uint8_t  temp_crc     = shtc3_rx_buff[2];
-//  uint8_t  hum_crc      = shtc3_rx_buff[5];
-//  uint8_t  temp_crc_act = crc8(shtc3_rx_buff, SHTC3_RX_MEASURE_BUFF_LEN, 0x31);
-//  uint8_t  hum_crc_act  = crc8(shtc3_rx_buff + 3, SHTC3_RX_MEASURE_BUFF_LEN, 0x31);
+  uint16_t temp_raw = (shtc3_rx_buff[0] << 8) | shtc3_rx_buff[1];
+  uint16_t hum_raw  = (shtc3_rx_buff[3] << 8) | shtc3_rx_buff[4];
+  //  uint8_t  temp_crc     = shtc3_rx_buff[2];
+  //  uint8_t  hum_crc      = shtc3_rx_buff[5];
+  //  uint8_t  temp_crc_act = crc8(shtc3_rx_buff, SHTC3_RX_MEASURE_BUFF_LEN,
+  //  0x31); uint8_t  hum_crc_act  = crc8(shtc3_rx_buff + 3,
+  //  SHTC3_RX_MEASURE_BUFF_LEN, 0x31);
   // Convert values and return them
-  *temperature = -45 + (175.0 * (float)temp_raw / (0x01 << 16));
-  *humidity    = 100 * (float)hum_raw / (0x01 << 16);
+  *temperature = -45 + (175.0 * (float)temp_raw / (float)(0x01 << 16));
+  *humidity    = 100.0 * (float)hum_raw / (float)(0x01 << 16);
   // TODO: Check CRC values
   return true;
 }
